Add basic support for languages style
This patch adds minimal support for changing the language/theme/style
seen in scintilla. All the information (?) that is available on the
theme is applied.

TODO:
seems like Notepad++ has a separated file which contains the language
model (whatever it means). I will have to match this to the style.